Finish RREF implementation and determine the maximum degrees of freedom
The maximum degrees of freedom in the regular input is 3. This can be easily permuted if the order of operations is otherwise known.

+++ day-10/src/test/java/io/github/stdonnelly/adventofcode/day10/mapper/MachineToMatrixMapperTest.java
@@ -46,7 +46,7 @@ class MachineToMatrixMapperTest {
         };
     final Matrix.Row expected =
         new Matrix.Row(
-            Arrays.stream(expectedRowString.split(":")).mapToInt(Integer::parseInt).toArray());
+            Arrays.stream(expectedRowString.split(":")).mapToDouble(Double::parseDouble).toArray());
 
     assertEquals(expected, MAPPER.mapRow(buttonSchematics, joltageRequirement, joltageIndex));
   }
